+++ tests/ui/cli/test_cli_summarize.py
@@ -186,6 +186,7 @@ def test_cli_summarize_custom_units() -> None:
     ]
 
 
+@pytest.mark.remote
 @pytest.mark.parametrize(
     "fmt",
     [